made args validation common
