+++ pvoutput/utils.py
@@ -1,9 +1,7 @@
 """ Utilities """
-import re
 from datetime import datetime
 from math import floor
 
-from .exceptions import DonationRequired, InvalidRegexpError
 
 BASE_URL = "https://pvoutput.org/service/r2/"
 
@@ -94,82 +92,3 @@ def get_rate_limit_header(response_object) -> dict:
         if key.startswith("X-Rate-Limit"):
             retval[key] = response_object.headers[key]
     return retval
-
-
-# pylint: disable=too-many-branches
-def validate_data(self, data, apiset):
-    """Does a super-simple validation based on the api def raises errors if it's wrong, returns True if it's OK
-
-    This'll only raise an error on the first error it finds
-
-    :param data: the data to validate.
-    :type data: dict
-
-    :param apiset: A set of validation rules, eg: pvoutput.ADDSTATUS_PARAMETERS
-    :type apiset: dict
-
-    :raises TypeError: if the type testing fails.
-    :raises ValueError: if you're trying to pass an invalid value.
-    :raises pvoutput.InvalidRegexpError: if value does not match the regexp in format.
-    """
-    # if you set a 'required_oneof' key in apiset, validation will require at least one of those keys to be set
-    if "required_oneof" in apiset.keys() and (
-        len([key for key in data.keys() if key in apiset["required_oneof"]["keys"]])
-        == 0
-    ):
-        raise ValueError(
-            f"one of {','.join(apiset['required_oneof']['keys'])} MUST be set"
-        )
-    for key in apiset:
-        # check that that required values are set
-        if apiset[key].get("required", False) and key not in data.keys():
-            raise ValueError(f"key {key} required in data")
-    # check there's no extras
-    for key in data.keys():
-        if key not in apiset.keys():
-            raise ValueError(f"key {key} isn't valid in the API spec")
-        if apiset[key].get("type", False) and not isinstance(
-            data[key], apiset[key]["type"]
-        ):
-            raise TypeError(
-                f"data[{key}] type ({type(data[key])} is invalid - should be {str(apiset[key]['type'])})"
-            )
-
-    for key in data:
-        format_string = apiset[key].get("format", False)
-        if format_string:
-            try:
-                compiled = re.compile(format_string)
-                match = compiled.match(data[key])
-                if match is None:
-                    raise ValueError(
-                        f"key '{key}', with value '{data[key]}' does not match '{format_string}'"
-                    )
-            except re.error as error:
-                raise InvalidRegexpError(
-                    f"Error for key '{key}' with format '{format_string}': {error}"
-                ) from error
-
-    # TODO: 'd' can't be more than 14 days ago, if a donator, goes out to 90
-    # check if donation_made == True and age of thing
-    # if self.donation_made:
-    #     # check if more than 90 days ago
-    # else:
-    #     # check if more than 14 days ago
-
-    # check for donation-only keys
-    if not self.donation_made:
-        donation_required_keys = [
-            key for key in data.keys() if apiset[key].get("donation_required", False)
-        ]
-        for key in data.keys():
-            if key in donation_required_keys:
-                raise DonationRequired(
-                    f"key {key} requires an account which has donated"
-                )
-    if int(data.get("v3", 0)) > 200000:
-        raise ValueError(f"v3 cannot be higher than 200000, is {data['v3']}")
-    if int(data.get("v4", 0)) > 100000:
-        raise ValueError(f"v4 cannot be higher than 100000, is {data['v4']}")
-
-    return True
